+++ packages/core/src/utils/index.ts
@@ -24,8 +24,6 @@ export {
   resolvePath,
   resolvePathString,
   resolveTags,
-  parseTransportPath,
-  isTransportPath,
   type TagMode,
   type TagOptions,
 } from "./path-utils";